Qos provider model updated

+++ src/main/java/sm/providing/learning/DeepQ.java
@@ -51,7 +51,7 @@ class DeepQ {
       boolean isValid = false;
       if (epsilon > rnd.nextDouble()) {
          while (!isValid) {
-            lastAction = rnd.nextInt(indArrayOutput.size(1));
+            lastAction = rnd.nextInt((int) indArrayOutput.size(1));
             if (actionMask[lastAction] == 1)
                isValid = true;
          }
